bug fixes+ comment adding+ tag adding
matan - couldn't find the function for adding tag and rate so i used the
regular function. please implement the function we talked about and
change it to match the one i used. The funciton I used is static so i
passed the current user id.
When trying to get comments form the user - i get an error form the db.
I attached a picture to the whatsup group. check it out.